minor documentation changes, a filter changed
--HG--
extra : convert_revision : svn%3A948f8045-528a-dd11-95b9-00e08143843c/projects/radio_data_system/trunk%40414

+++ src/lib/gr_rds_data_encoder.h
@@ -83,7 +83,7 @@ private:
 	int d_buffer_current;
 	int d_sign_last;
 	int d_zero_cross;
-	int d_current_out;
+	float d_current_out;
 
 // Functions
 	friend gr_rds_data_encoder_sptr gr_rds_make_data_encoder (const char *xmlfile);
